Add context
First sight properly working
