+++ src/components/TodoSearch.js
@@ -1,6 +1,8 @@
-import React from 'react'
+import React, { useContext } from 'react'
+import { TodoContext } from '../TodoContext/TodoContex';
 
-function TodoSearch({ searchValue, setSearchValue }) {
+function TodoSearch() {
+  const { searchValue, setSearchValue } = useContext(TodoContext);
 
   const onSearchValueChange = (event) => {
     setSearchValue(event.target.value);
